producer/requirements.txt
kafka-python==2.0.1
protobuf3==0.2.1